Issue #2 - NAT Rule Creation Error
When attempting to create a NAT rule it was erroring out. The issue was
coming from the variable being defined as whether the rule was immediate
or not. When this variable was being defined for a NAT rule it generated
the error as seen in issue #2. This has been resolved now.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>